Backend/requirements.txt
flask
flask-cors
ultralytics
torch
torchvision
opencv-python-headless
pillow
requests
gunicorn
